(*) Better Performance and new score items

+++ UF1_Proyecto/app/src/main/java/com/example/uf1_proyecto/Activity/ProfileFragment.kt
@@ -2,11 +2,13 @@ package com.example.uf1_proyecto.Activity
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.*
+import android.widget.Button
+import android.widget.TextView
+import android.widget.Toast
+import androidx.fragment.app.Fragment
 import com.example.uf1_proyecto.R
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
@@ -41,14 +43,24 @@ class ProfileFragment : Fragment() {
         val intent = Intent(requireContext(), IntroActivity::class.java)
 
         borrarBtn.setOnClickListener {
-            FirebaseAuth.getInstance().currentUser!!.delete()
-            Toast.makeText(requireContext(), "Borrado", Toast.LENGTH_SHORT).show()
-            startActivity(intent)
+            val alertDialog: android.app.AlertDialog.Builder = android.app.AlertDialog.Builder(requireContext())
+            alertDialog.setTitle("Borrar cuenta")
+            alertDialog.setMessage("¿Estás seguro de que quieres borrar tu cuenta?")
+            alertDialog.setPositiveButton("Sí") { _, _ ->
+                FirebaseAuth.getInstance().currentUser!!.delete()
+                Toast.makeText(requireContext(), "Borrado", Toast.LENGTH_SHORT).show()
+                startActivity(intent)
+            }
         }
         cerrarSesionBtn.setOnClickListener {
-            FirebaseAuth.getInstance().signOut()
-            Toast.makeText(requireContext(), "Cerrando sesión", Toast.LENGTH_SHORT).show()
-            startActivity(intent)
+            val alertDialog: android.app.AlertDialog.Builder = android.app.AlertDialog.Builder(requireContext())
+            alertDialog.setTitle("Cerrar sesión")
+            alertDialog.setMessage("¿Estás seguro de que quieres cerrar sesión?")
+            alertDialog.setPositiveButton("Sí") { _, _ ->
+                FirebaseAuth.getInstance().signOut()
+                Toast.makeText(requireContext(), "Cerrando sesión", Toast.LENGTH_SHORT).show()
+                startActivity(intent)
+            }
         }
     }
